LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/revert-val-006-expected.txt


FAIL The revert keyword works with @keyframes assert_equals: expected "21.440000534057617px" but got "0px"
